Begin porting to the more up-to-date BSD sources.

+++ src/kern/ext2_apple.c
@@ -39,18 +39,19 @@ static const char whatid[] __attribute__ ((unused)) =
 #include <sys/sysctl.h>
 
 #include "ext2_apple.h"
-#include <gnu/ext2fs/ext2_fs.h>
-#include <gnu/ext2fs/ext2_fs_sb.h>
-#include <gnu/ext2fs/inode.h>
-#include <gnu/ext2fs/fs.h>
-#include <gnu/ext2fs/ext2_extern.h>
+#include <fs/ext2fs/fs.h>
+#include <fs/ext2fs/inode.h>
+#include <fs/ext2fs/ext2fs.h>
+#include <fs/ext2fs/inode.h>
+#include <fs/ext2fs/fs.h>
+#include <fs/ext2fs/ext2_extern.h>
 
 // missing clib functions (as of 10.4.x)
 __private_extern__ char*
 e_strrchr(const char *s, int c)
 {
    char *name;
-   int i;
+   size_t i;
    
    name = NULL;
    i = strlen(s);
